+++ containers/addresses/missingKeys/MissingKeysStatus.tsx
@@ -1,35 +1,29 @@
 import { c } from 'ttag';
 import React from 'react';
-import { Badge, LoaderIcon } from '../../index';
-
-export enum STATUS {
-    QUEUED,
-    DONE,
-    FAILURE,
-    LOADING
-}
+import { Badge, LoaderIcon } from '../../../index';
+import { Status } from './interface';
 
 interface Props {
-    type: STATUS;
+    type: Status;
     tooltip?: string;
 }
-const Status = ({ type, tooltip }: Props) => {
-    if (type === STATUS.QUEUED) {
+const MissingKeysStatus = ({ type, tooltip }: Props) => {
+    if (type === Status.QUEUED) {
         return <Badge type="default">{c('Info').t`Queued`}</Badge>;
     }
 
-    if (type === STATUS.DONE) {
+    if (type === Status.DONE) {
         return <Badge type="success">{c('Info').t`Done`}</Badge>;
     }
 
-    if (type === STATUS.FAILURE) {
+    if (type === Status.FAILURE) {
         return <Badge type="error" tooltip={tooltip}>{c('Error').t`Error`}</Badge>;
     }
 
-    if (type === STATUS.LOADING) {
+    if (type === Status.LOADING) {
         return <LoaderIcon />;
     }
     return null;
 };
 
-export default Status;
+export default MissingKeysStatus;
