+++ containers/addresses/missingKeys/interface.ts
@@ -1,0 +1,17 @@
+import { Address } from 'proton-shared/lib/interfaces/Address';
+
+export enum Status {
+    QUEUED,
+    DONE,
+    FAILURE,
+    LOADING
+}
+
+export interface AddressWithStatus extends Address {
+    status: {
+        type: Status;
+        tooltip?: string;
+    };
+}
+
+export type SetFormattedAddresses = React.Dispatch<React.SetStateAction<AddressWithStatus[]>>;
